:recycle: Refactor common llms-txt generation code to xy-shared.
