+++ packages/xy-shared/package.json
@@ -10,6 +10,7 @@
     "generate:component": "turbo gen react-component"
   },
   "devDependencies": {
+    "@types/mdast": "^4.0.4",
     "@types/node": "^22.14.1",
     "@types/react": "^19.2.0",
     "@types/react-dom": "^19.2.0",
@@ -48,9 +49,15 @@
     "next": "^14.2.15",
     "nextra": "4.3.0-alpha.29",
     "react-lite-youtube-embed": "^2.5.1",
+    "remark-frontmatter": "^5.0.0",
+    "remark-gfm": "^4.0.1",
+    "remark-mdx": "^3.1.1",
+    "remark-parse": "^11.0.0",
+    "remark-stringify": "^11.0.0",
     "tailwindcss": "^3.4.14",
     "tailwindcss-animate": "^1.0.7",
     "timeago-react": "^3.0.7",
+    "unified": "^11.0.5",
     "use-error-boundary": "^2.0.6"
   },
   "peerDependencies": {
