+++ packages/xy-shared/server/index.ts
@@ -7,3 +7,4 @@ export * from './timeline-event';
 export * from './remote-code-viewer';
 export * from './toc';
 export * from './example-utils';
+export * from './llms-txt';
