+++ sites/reactflow.dev/package.json
@@ -30,21 +30,14 @@
     "react": "^19.2.0",
     "react-dom": "^19.2.0",
     "react-icons": "^5.5.0",
-    "remark-frontmatter": "^5.0.0",
-    "remark-gfm": "^4.0.0",
-    "remark-mdx": "^3.1.0",
-    "remark-parse": "^11.0.0",
-    "remark-stringify": "^11.0.0",
     "timeago-react": "^3.0.7",
     "ui-components": "workspace:*",
-    "unified": "^11.0.5",
     "xy-shared": "workspace:*"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.3.1",
     "@notionhq/client": "^2.3.0",
     "@types/d3": "^7.4.3",
-    "@types/mdast": "^4.0.4",
     "@types/node": "^22.14.1",
     "@types/react": "^19.2.0",
     "eslint-config-xyflow": "workspace:*",
