+++ sites/reactflow.dev/src/app/(llms-txt)/llms-medium.txt/route.tsx
@@ -1,15 +1,11 @@
 // ROUTE -----------------------------------------------------------------------
 
-import { buildLLMSTxt, ALL_SECTIONS } from '../utils';
+import { buildLLMSTxt } from '../utils';
 
 export const dynamic = 'force-static';
 
-// Exclude the examples section from the full LLMSTxt
-const sections = { ...ALL_SECTIONS };
-delete sections.examples;
-
 export async function GET() {
-  const body = await buildLLMSTxt(sections);
+  const body = await buildLLMSTxt(['learn', 'ui']);
 
   return new Response(body, {
     status: 200,
