+++ sites/svelteflow.dev/package.json
@@ -23,11 +23,6 @@
     "fast-xml-parser": "^5.2.1",
     "next": "^15.5.7",
     "next-sitemap": "^4.2.3",
-    "remark-frontmatter": "^5.0.0",
-    "remark-gfm": "^4.0.0",
-    "remark-parse": "^11.0.0",
-    "remark-stringify": "^11.0.0",
-    "unified": "^11.0.5",
     "nextra": "4.5.1",
     "nextra-theme-docs": "4.5.1",
     "react": "^19.2.0",
@@ -37,7 +32,6 @@
   },
   "devDependencies": {
     "@notionhq/client": "^2.3.0",
-    "@types/mdast": "^4.0.4",
     "@types/node": "^22.14.1",
     "@types/react": "^19.2.0",
     "eslint-config-xyflow": "workspace:*",
